+++ compose-app/src/main/java/com/mapbox/maps/compose/testapp/examples/animation/MapViewportAnimationsActivity.kt
@@ -104,6 +104,7 @@ public class MapViewportAnimationsActivity : ComponentActivity() {
                           )
                         )
                       )
+                      .animationDurationMs(600L)
                       .padding(EdgeInsets(100.0, 100.0, 100.0, 100.0))
                       .build(),
                     defaultTransitionOptions = DefaultViewportTransitionOptions.Builder()